+++ cmd/parse/parse.go
@@ -3,12 +3,9 @@ package parse
 import (
 	parseauth "github.com/forbole/bdjuno/v4/cmd/parse/auth"
 	parsebank "github.com/forbole/bdjuno/v4/cmd/parse/bank"
-	parsedistribution "github.com/forbole/bdjuno/v4/cmd/parse/distribution"
+	parseccv "github.com/forbole/bdjuno/v4/cmd/parse/ccv"
 	parsefeegrant "github.com/forbole/bdjuno/v4/cmd/parse/feegrant"
-	parsegov "github.com/forbole/bdjuno/v4/cmd/parse/gov"
-	parsemint "github.com/forbole/bdjuno/v4/cmd/parse/mint"
 	parsepricefeed "github.com/forbole/bdjuno/v4/cmd/parse/pricefeed"
-	parsestaking "github.com/forbole/bdjuno/v4/cmd/parse/staking"
 	parseblocks "github.com/forbole/juno/v4/cmd/parse/blocks"
 	parsegenesis "github.com/forbole/juno/v4/cmd/parse/genesis"
 	parsetransaction "github.com/forbole/juno/v4/cmd/parse/transactions"
@@ -28,13 +25,10 @@ func NewParseCmd(parseCfg *parse.Config) *cobra.Command {
 		parseauth.NewAuthCmd(parseCfg),
 		parsebank.NewBankCmd(parseCfg),
 		parseblocks.NewBlocksCmd(parseCfg),
-		parsedistribution.NewDistributionCmd(parseCfg),
+		parseccv.NewCcvCmd(parseCfg),
 		parsefeegrant.NewFeegrantCmd(parseCfg),
 		parsegenesis.NewGenesisCmd(parseCfg),
-		parsegov.NewGovCmd(parseCfg),
-		parsemint.NewMintCmd(parseCfg),
 		parsepricefeed.NewPricefeedCmd(parseCfg),
-		parsestaking.NewStakingCmd(parseCfg),
 		parsetransaction.NewTransactionsCmd(parseCfg),
 	)