+++ modules/actions/types/utils.go
@@ -1,5 +1,0 @@
-package types
-
-type GraphQLError struct {
-	Message string `json:"message"`
-}