+++ modules/types/sources.go
@@ -2,52 +2,28 @@ package types
 
 import (
 	"fmt"
-	"os"
-
-	"github.com/cosmos/cosmos-sdk/simapp"
-	"github.com/tendermint/tendermint/libs/log"
 
 	"github.com/cosmos/cosmos-sdk/simapp/params"
 	"github.com/forbole/juno/v4/node/remote"
 
+	wasmtypes "github.com/CosmWasm/wasmd/x/wasm/types"
 	banktypes "github.com/cosmos/cosmos-sdk/x/bank/types"
-	distrtypes "github.com/cosmos/cosmos-sdk/x/distribution/types"
-	govtypes "github.com/cosmos/cosmos-sdk/x/gov/types"
-	minttypes "github.com/cosmos/cosmos-sdk/x/mint/types"
 	slashingtypes "github.com/cosmos/cosmos-sdk/x/slashing/types"
-	stakingkeeper "github.com/cosmos/cosmos-sdk/x/staking/keeper"
-	stakingtypes "github.com/cosmos/cosmos-sdk/x/staking/types"
+	remotewasmsource "github.com/forbole/bdjuno/v4/modules/wasm/source/remote"
 	"github.com/forbole/juno/v4/node/local"
 
-	nodeconfig "github.com/forbole/juno/v4/node/config"
-
 	banksource "github.com/forbole/bdjuno/v4/modules/bank/source"
-	localbanksource "github.com/forbole/bdjuno/v4/modules/bank/source/local"
 	remotebanksource "github.com/forbole/bdjuno/v4/modules/bank/source/remote"
-	distrsource "github.com/forbole/bdjuno/v4/modules/distribution/source"
-	localdistrsource "github.com/forbole/bdjuno/v4/modules/distribution/source/local"
-	remotedistrsource "github.com/forbole/bdjuno/v4/modules/distribution/source/remote"
-	govsource "github.com/forbole/bdjuno/v4/modules/gov/source"
-	localgovsource "github.com/forbole/bdjuno/v4/modules/gov/source/local"
-	remotegovsource "github.com/forbole/bdjuno/v4/modules/gov/source/remote"
-	mintsource "github.com/forbole/bdjuno/v4/modules/mint/source"
-	localmintsource "github.com/forbole/bdjuno/v4/modules/mint/source/local"
-	remotemintsource "github.com/forbole/bdjuno/v4/modules/mint/source/remote"
 	slashingsource "github.com/forbole/bdjuno/v4/modules/slashing/source"
-	localslashingsource "github.com/forbole/bdjuno/v4/modules/slashing/source/local"
 	remoteslashingsource "github.com/forbole/bdjuno/v4/modules/slashing/source/remote"
-	stakingsource "github.com/forbole/bdjuno/v4/modules/staking/source"
-	localstakingsource "github.com/forbole/bdjuno/v4/modules/staking/source/local"
-	remotestakingsource "github.com/forbole/bdjuno/v4/modules/staking/source/remote"
+	wasmsource "github.com/forbole/bdjuno/v4/modules/wasm/source"
+	nodeconfig "github.com/forbole/juno/v4/node/config"
 )
 
 type Sources struct {
 	BankSource     banksource.Source
-	DistrSource    distrsource.Source
-	GovSource      govsource.Source
-	MintSource     mintsource.Source
 	SlashingSource slashingsource.Source
-	StakingSource  stakingsource.Source
+	WasmSource     wasmsource.Source
 }
 
 func BuildSources(nodeCfg nodeconfig.Config, encodingConfig *params.EncodingConfig) (*Sources, error) {
@@ -55,55 +31,10 @@ func BuildSources(nodeCfg nodeconfig.Config, encodingConfig *params.EncodingConf
 	case *remote.Details:
 		return buildRemoteSources(cfg)
 	case *local.Details:
-		return buildLocalSources(cfg, encodingConfig)
-
+		return nil, fmt.Errorf("local source is not supported: %T", cfg)
 	default:
 		return nil, fmt.Errorf("invalid configuration type: %T", cfg)
 	}
-}
-
-func buildLocalSources(cfg *local.Details, encodingConfig *params.EncodingConfig) (*Sources, error) {
-	source, err := local.NewSource(cfg.Home, encodingConfig)
-	if err != nil {
-		return nil, err
-	}
-
-	app := simapp.NewSimApp(
-		log.NewTMLogger(log.NewSyncWriter(os.Stdout)), source.StoreDB, nil, true, map[int64]bool{},
-		cfg.Home, 0, simapp.MakeTestEncodingConfig(), simapp.EmptyAppOptions{},
-	)
-
-	sources := &Sources{
-		BankSource:     localbanksource.NewSource(source, banktypes.QueryServer(app.BankKeeper)),
-		DistrSource:    localdistrsource.NewSource(source, distrtypes.QueryServer(app.DistrKeeper)),
-		GovSource:      localgovsource.NewSource(source, govtypes.QueryServer(app.GovKeeper)),
-		MintSource:     localmintsource.NewSource(source, minttypes.QueryServer(app.MintKeeper)),
-		SlashingSource: localslashingsource.NewSource(source, slashingtypes.QueryServer(app.SlashingKeeper)),
-		StakingSource:  localstakingsource.NewSource(source, stakingkeeper.Querier{Keeper: app.StakingKeeper}),
-	}
-
-	// Mount and initialize the stores
-	err = source.MountKVStores(app, "keys")
-	if err != nil {
-		return nil, err
-	}
-
-	err = source.MountTransientStores(app, "tkeys")
-	if err != nil {
-		return nil, err
-	}
-
-	err = source.MountMemoryStores(app, "memKeys")
-	if err != nil {
-		return nil, err
-	}
-
-	err = source.InitStores()
-	if err != nil {
-		return nil, err
-	}
-
-	return sources, nil
 }
 
 func buildRemoteSources(cfg *remote.Details) (*Sources, error) {
@@ -114,10 +45,7 @@ func buildRemoteSources(cfg *remote.Details) (*Sources, error) {
 
 	return &Sources{
 		BankSource:     remotebanksource.NewSource(source, banktypes.NewQueryClient(source.GrpcConn)),
-		DistrSource:    remotedistrsource.NewSource(source, distrtypes.NewQueryClient(source.GrpcConn)),
-		GovSource:      remotegovsource.NewSource(source, govtypes.NewQueryClient(source.GrpcConn)),
-		MintSource:     remotemintsource.NewSource(source, minttypes.NewQueryClient(source.GrpcConn)),
 		SlashingSource: remoteslashingsource.NewSource(source, slashingtypes.NewQueryClient(source.GrpcConn)),
-		StakingSource:  remotestakingsource.NewSource(source, stakingtypes.NewQueryClient(source.GrpcConn)),
+		WasmSource:     remotewasmsource.NewSource(source, wasmtypes.NewQueryClient(source.GrpcConn)),
 	}, nil
 }